separate game class from index.js file
- Add style class
