+++ package.json
@@ -18,14 +18,19 @@
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
     "copy-webpack-plugin": "^8.1.0",
+    "css-loader": "^5.2.0",
     "jest": "^26.6.3",
     "jest-fetch-mock": "^3.0.3",
     "jest-localstorage-mock": "^2.4.8",
+    "style-loader": "^2.0.0",
     "webpack": "^5.28.0",
     "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2"
   },
   "dependencies": {
-    "phaser": "^3.54.0"
+    "bootstrap": "^4.6.0",
+    "jquery": "^3.6.0",
+    "phaser": "^3.54.0",
+    "popper.js": "^1.16.1"
   }
 }
