+++ src/index.js
@@ -1,31 +1,8 @@
 import 'phaser';
+import './style.css';
+import 'bootstrap';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
-import config from './Config/config';
-import GameScene from './Scenes/GameScene';
-import BootScene from './Scenes/BootScene';
-import PreloaderScene from './Scenes/PreloaderScene';
-import TitleScene from './Scenes/TitleScene';
-import OptionsScene from './Scenes/OptionsScene';
-import CreditsScene from './Scenes/CreditsScene';
-import Model from './Model';
 
-class Game extends Phaser.Game {
-  constructor(){
-    super(config);
-
-    // call Model class
-    const model = new Model();
-    this.globals = { model, bgMusic: null };
-
-    // Call Scenes
-    this.scene.add('Boot', BootScene);
-    this.scene.add('Preloader', PreloaderScene);
-    this.scene.add('Title', TitleScene);
-    this.scene.add('Options', OptionsScene);
-    this.scene.add('Credits', CreditsScene);
-    this.scene.add('Game', GameScene);
-    this.scene.start('Boot');
-  }
-}
 
 window.game = new Game();
