+++ src/style.css
@@ -1,0 +1,8 @@
+.img-banner {
+  width: 100%;
+  height: 50vh;
+}
+
+.logo {
+  width: 100%
+}
